fix - Update tax rate methods to return BigDecimal instead of double for consistency

+++ src/main/java/co/invest72/investment/domain/tax/FixedTaxRate.java
@@ -41,8 +41,8 @@ public class FixedTaxRate implements TaxRate {
 	}
 
 	@Override
-	public double getValue() {
-		return value.doubleValue();
+	public BigDecimal getValue() {
+		return value;
 	}
 
 	@Override
