+++ src/main/java/co/invest72/investment/domain/tax/TaxBenefit.java
@@ -30,6 +30,6 @@ public class TaxBenefit implements Taxable {
 
 	@Override
 	public double getTaxRate() {
-		return taxRate.getValue();
+		return taxRate.getValue().doubleValue();
 	}
 }
